Update project 11 for iOS 17 and SwiftData

+++ Project 11 - Bookworm/Bookworm/Bookworm/Commons/Views/Components/EmojiRatingView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct EmojiRatingView: View {
-    let rating: Int16
+    let rating: Int
 
     var body: some View {
         switch rating {
